Update workload control objects.
The TSEM kernel driver was updated in order to provide a pathway
for decreasing the complexity of the userspace trust
orchestrators.  This change involve the embedding of the security
modeling namespace context in the orchestrator process.  That
provides direct access to either the events being exported or the
charcteristics fo the model that is running in the kernel.

In the previous architecture there was a need to have a monitor
process running in the context of the security modeling namespace
that had access to the characteristics of the namespace.

In the case of internally modeled namespaces this was causing a
'Heisenberg' condition where interrogation of the characteristics
of the mdoeling namespace resulted in a change in the state of
the model.  If the modeling namespace did not a connection from
the orchestration console an attempt to interrogate the namespace
would generate a forensics event.

The externally modeled namespaces had a complex architecture that
used a pipe between the workload monitor and the launch process
where the model was being run.  Exported events had to be sent
over this pipe which necessitated extra context switches for
every event.

The other major issue with this architecture is that an extra process was
required for each namespace that was running.

This commit updates the TSEMevent and TSEMworkload objects to
support this new model.

Previously, the TSEMevent object read events from the pipe that
was attaching the modeling process to the workload monitor
running in the context of the namespace.  The object was updated
to add the TSEMevent->read_export method that handles read from
the event export control plane file.

This method reads a single event from the control plane file and
uses a pointer passed to it to signal to the caller that a read
had ended in in a ENODATA condition, ie. there were no longer any
events in the queue to read.

There were major changes in the TSEMworkload object that
implemented the necessary infrastructure for running workloads in
this new model

The function signature for the TSEMworkload->run_monitor method
was modified to drop the need to pass the PID of the workload
process into the monitor.  The trust orchestrator calls the
TSEMworkload->run_workload method first when setting up a
security modeling namespace so the PID value of the workload is
known internally by the object.

There were significant modifictions to the
TSEMworkload->run_workload method in order to support this
architecture.

One of the notable changes was for a pipe to be created that
connects the workload monitor process with the forked workload
process.  The workload writes the container identifier over this
pipe which provides the workload monitor with the container
identifier and signals the workload monitor that the workload
is starting execution.

The _setup_namespace() helper function was updated to write the
identifier over the pipe after the namespace is created.

The _open_event_file() function was implemented to provide
support for the workload monitor to open the export file.

The _process_events() function was implemented to encapsulate the
processing of events that are exported.  That process will run in
a loop reading event descriptions and passing them to the event
processing function until the control plane returns an ENODATA
value indicating that there are no further events to be
processed.

All of this infrastructure is still very early in its
development.  Testing has indicated that there might be a
possibility of the trust orchestration stack stalling when the
workload process terminates.  Additional testing will be needed
to verify that there are no subtle issues with this.