Add support to TSEMworkload for root export.
The TSEMworkload->set_root_mode method was added in order to
allow the TSEMworkload object to be used for exporting the
security events from the root modeling namespace.

This method opens the control plane export file for the root
modeling namespace and returns that file descriptor to the caller
via a pointer passed to the method.