Update quixote-export to use new TSEMworkload infrastructure.
This commit converts the trust orchestrator that handles simple
export of security event descriptions to use enhanced
implementation of the TSEMworkload object.

The enhancements to the TSEMworkload object significantly
simplifies the implementation of the run_workload() and
run_broker_workload() functions.

The majority of the changes in this commit involved updating the
support for exporting events from the root security modeling
namespace.

The Gaggle object that holds the queue of event descriptions to
be exported from the root modeling namespace was moved to static
scope in order to avoid having to pass it into the various helper
functions.

Rather than an open coded implementation of the code that reads
the events exported from the root modeling namespace a TSEMevent
object is used.

An instance of this object is implemented in the export_root()
function to read the events that were currently being held in the
kernel.  In 'follow' mode the TSEMworkload->run_monitor is used
to poll for and read the events that are then pessed to the event
handler for the root modeling namespace.

The _queue_event() function was implemented to allow the Gaggle
of events to be updated using a TSEMevent object.

The _process_event() function was implemented that is the event
handler that is invoked by the TSEMworkload->run_manager object.
This function uses the _queue_event() to queue the event and then
checks for the need to flush the queue.

The export_root() function was updated to take the TSEMworkload
object that is opened for the broker and file based exports.