Support internally modeled namespaces.
The TSEMworkload->configure_internal method was added in order to
support internally modeled security namespaces.

Since internally modeled namespaces do not have an export file
the type of the object is checked in the _open_event_file()
function so the function can be bypassed for internally modeled
namespaces.

Since internally modeled namespaces do not need to poll on an
event fill the polling information structure is memset to all
nulls in order to prevent the poll command from generating a
fault due to stack garbage being in the poll_data array.