Manage internally modeled namespaces with TSEMworkload.
This commit introduces the substantial number of changes that are
needed in order to support internally modeled namespaces in the
new architectural model where the trust orchestrator has
posession of the security model context.

Despite the number and size of the changes the conversion was
reasonably straight forward at this current level of maturity.  A
large amount of code was dropped since a lot of the open-coded
functionality is now handled by the TSEMworkload object.

The fire_cartridge() function was replaced with the
run_workload() function in order to maintain structural
similarity with the quixote-us orchestrator.

Loading of security models is currently not functional at this
commit.