Implement control over release of workload process.
One of the issues that needs to be addressed in order to have
loadable security model support for internally modeled namespaces
is to have the ability of the trust orchestrator to release the
workload when the model is loaded and ready to control the
workload.

The event_pipe member of the TSEMworkload state structure was
re-purposed to be used as a pipe that connects the trust
orchestrator to the workload process.

After the namespace is created the workload process writes the
namespace identifier over the pipe to the trust orchestrator.
This serves as a gating event for the trust orchestrator to
continue setup for the workload.

After writing the container identifier the workload issues a one
byte read over the pipe in order to block the process while the
the trust orchestrator completes its setup.  When the
orchestrator is ready to begin polling for events it writes one
byte over the pipe in order to release the workload that then
continues onward to execute the workload.

All of the work to setup the namespace was migrated from the
TSEMworkload->run_workload method to the _setup_namespace helper
method.