Update HurdLib to support access to File errors.
The TSEM kernel implementation was updated to return ENOENT if
an attempt to open the /sys/kernel/security/tsem/id control plane
file was called without the TMA context of the calling process
having been set.

This was implemented in order to allow the orchestrator process
to 'poll' for whether or not the creation of the security
modeling namespace had completed.

The error code from the last failed I/O operation was being
preserved but there was no API to access it through the object.
Implement the File->error method in order to return the error
value.