Update TSEMworkload implementation.
Properly handling the configuration of an internaly modeled
namespace ended up leading to a major update to the TSEMworkload
object.

This update also addresses intermittent startup failures that
were caused by a lack of synchronization between the orchestrator
process and the workload.

This update eliminated the need to have separate
TSEMworkoad->run_workload and TSEMworkload->run_monitor methods.
The start of the workload is now initiated by what was formely
the TSEMworkload->run_monitor method and which is referred to, as
of this commit, as the TSEMworkload->run_workload method.

A String object was added to the state of the TSEMworkload object
in order to store a security model/map if one was defined for the
workload.  The TSEMworkload->configure_external and
TSEMworkload->configure_internal methods were implemented to take
a character pointer to the pathname of the file containing the
security model.

The following three helper methods were implemented in order to
assist with the creation and execution of the workload:

_create_namespace()
_configure_namespace()
_start_workload()

The _create_namespace helper is called by the workload to create
the namespace and to wait until the orchestrator process has
setup the namespace.

The _configure_namespace helper is called by the orchestrator
process to configure the namespace.  In the case of an internally
modeled namespace this includes loading the security model into
the namespace.  The function blocks until an open of the
/sys/kernel/security/tsem/id file succeeds.  This indicates that
the security modeling namespace has been configured and is ready
to begin executing the workload.

The start of the workload is signaled by the orchestrator
processing writing a single byte over a pipe that connects the
orchestrator process to the workload.

After receving the release signal the _start_workload() process
interprets the characteristics of the workload and carries out
the necessary actions to run the process that will implement the
workload.

This commit also significantly changes the monitor loop that
polls for events originating from the kernel export file or the
management interface.

The file descriptor member of each poll structure is set to a -1
in the case where the member of the poll is not being polled.
This happens, for example, in an internally modeled namespace
where there is no need to read an export file.