Update orchestrators to use new TSEMworkload implementation.
This commit updates the orchestrators that implement in kernel
modeling, userspace modeling and export of events to use the new
TSEMworkload implementation.

This largely consisted of changing separate calls to the
TSEMworkload->run_workload and TSEMworkload->run_monitor methods
to a single call to the TSEMworkload->run_workload method, that
now takes the arguments that were previously supplied to the old
TSEMworkload->run_monitor method.

The load_model() function in the quixote.c implementation was
modified to take a String argument rather than a pointer to a
null-terminated character array.  This is secondary to the fact
that the TSEMworkload method calls the function that is
responsible for loading the model and passes to it the String
object that was generated when the workload was configured.