Provide access to control object for internally modeled namespaces.
An internally modeled namespace needs significantly more access
to the TSEMcontrol object than do externally modeled or export
namespaces.

Rather than providing inheritance wrappers for the functionality
a pointer is provided to the TSEMworkload->configure_internal
method in order to allow that method to enable the caller to have
access to the TSEMcontrol object.

This could be argued to be a layering violation but it doesn't
burden the TSEMworkload object with function slots that would be
needed only for internally modeled namespaces.

This imposes a requirement that any file that includes the
TSEMworkload header also needs to include the TSEMcontrol header.
This include was added to the quixote-export utility to avoid
warning messages.