Delete path description if address is in use.
If the pathname to a UNIX domain socket alread exists the call to
bind to the socket will fail with an address in use error
message.

The destructor (->method) for the LocalDuct object will check to
see if a pathname for the object has been created, and if so,
will unlink the path.

This behavior, unfortunately, causes an existing UNIX domain
socket path to be deleted if an attempt is made to open a second
socket with the same name/address.

Check the return code for the bind call and it it returns
EADDRINUSE destroy the description to the path so that the
subsequent and expected destruction of the object will not unlink
a path that it should not have possession of.