Integrate workload managent into TSEMworkload object.
This commit represents a significant structural change to the
orchestrator infrastructure by moving the setup of the workload
management infrastructure into the TSEMworkload object itself.
Previously there had been a requirement to pass the LocalDuct
object that would be used to manage the workload into the object
through the TSEMworkload->run_workload method.

This removes the need to implement boilerplate code in each trust
orchestrator and enables the failure to setup management
infrastructure to be handled in a manner that does not leave
orphaned workload processes present.

This change also lays the groundwork for using a single name for
a workload, regardless of whether it is container, process or
export based.

The following methods were updated to take as an argument a
pointer to a null-terminated character buffer containing the name
of the workload:

TSEMworkload->configure_export
TSEMworkload->configure_internal
TSMEworkload->configure_external

This pointer is converted into a String object that is maintained
in the workload state until the workload object is destroyed.

This character array is used to construct the bundle string for
container based workloads so the passing of the workload name to
the TSEMworkload->set_container method was dropped.

Since the orchestratror process may have multiple reasons to stop
the workload the writing of the control signal over the pipe
connecting the orchestrator and workload process was consolidated
into the following internal method:

TSEMworkload->_signal_workload

A boolean value is passed to this method to indicate whether the
workload should be terminated after the namespace was created or
allowed to execute the desired workload.

The _signal_workload method sends what a boolean value across the
pipe to specify whether or not the workload should be started or
terminated.

If the signal is to stop the workload a SIGKILL signal is sent to
the workload since it cannot terminate without any security
events that it generates being processed.  Given that there is no
desire to start the workload it is unconditionally killed in
order to simplify termination handling.