Update orchestration infrasture to use new workload model.
The TSEMworkload object was updated to implement management of
the UNIX domain based control plane.  This removes the need to
have each orchestrator create a LocalDuct object and pass it into
the method that initiates the workload.

In addition to this change a new model was adopted where each
workload is expected to have a name that it is to be known by.
The orchestrators were updated so that it is expected that the -w
argument will be used to specify a mandatory workload name.

Previously the orchestration infrastructure had used separate
directories for container vs. process based workloads.  Since the
infrastructure no longer considers them to be different all of
the management sockets are now created in the following
directory:

/var/lib/Quixote/mgmt

Moving forward the objective will be to consolidate all of the
management activities into a dedicated object since it is
anticipated that network based management will also be
implemented in addition to local domain sockets.