Allow command-line output of all model parameters.
This update reflects a major re-structuring of the
quixote-console utility in order to increase its ability to be
used for constructing security models.

The previous implementation of the utility was limited with
respect to being able to output the security coefficients that
were associated with forensics events.  This limited the ability
to generate updates to a model from security violations.

In 'onehost' mode there is now a mode designation that specifies
what type of data is to be output, for example trajectory or
forensics data.

The type designation now specifies, within a mode type, what type
of output is to be generated.  There are three separate types of
output:

event description
event counts
event coefficients

The following command-line options select the type of output:

-E	Event logs.
-F	Forensics information.
-M	Security model.
-S	Security state.
-T	Execution information.

By default security event descriptions are selected as the type
of output.

The following switches select alternate types of output:

-c	Count of events.
-s	Security state coefficients.

The following command-line switch was added in order to
facilitate the ability to fold forensic coefficients back into a
security state model:

-p

The -p switch specifies that the 'state ' prefix should be
prepended to each security state coefficient so the output can be
inserted directly into an existing security model map.