Interpret log events in the context of the model.
The TSEM kernel driver was updated to export descriptions of
security events that are invoked by an untrusted process rather
than a log entry describing the event.

The TSEMevent->extract_export method was updated to add the
event{} description structure when a TSEM_EVENT_LOG export event
was encountered.

The add_log() function was renamed to add_log_event().  The
security event description is parsed and the current Model object
is updated with the event description.  If the event is
inconsistent with the model being implemented the event will be
retained as a forensics event.  This allows these events to be
easily re-integrated into the security model.