Simply development of an updated security model definition.
The commit updates the quixote-console utility to simplify the
creation of an updated security model based on forensic
violations from a previous model.

The -u command-line argument was added to modify the behavior of
a -M mode invocation to also include any forensic coefficients
that may have been generated in a security modeling namespace.

For example, the following command invocation:

quixote-console -M -u

Will output a security model based on all of the security state
coefficients from the current namespace along with any security
coefficients that have been associated with forensic events.

A comment is placed in the model file that is generated that
indicates the number of additional state coefficients that were
added.  The comment also serves to delimit those coefficients
from the normal model coefficients.