Implement event timestamping.
Elasticsearch has the ability to do time series analysis if there
is timestamp information on a set of documents added to an index.
This commit adds the ability for the quixote-export trust
orchstrator to add timestamps to each event that is exported.

The -a command-line option was added to the quixote-export tool
to enable tha addition of timestamps.

The add_timestamp() function was added to encapsulate the
addition of the timestamp field to an event description.  The
export{} structure was modified so that it has the following
form:

export: {"type": "EVENT_TYPE", "@timestamp": "YYYY-MM-DDTHH:MM:SS.NNN"}

Where NNN is a nine digit nanosecond timestamp.

The remainder of the exported event description remains the same.