Implement Elasticsearch client object.
The ESclient object is implemented in this commit.  This object
is designed to provide an API object for issueing requests to an
Elasticsearch (ES) instance.  The libcurl library is used to
implement the protocol implementation for the

In this initial implemented the following methods, in addition to
the destructor are implemented:

ESclient->init
ESclient->inject

The ->init method is called in order to setup the object for
communication with a specific index on an ES instance.

This method takes as arguments the host to connect to, the index
that will be manipulated and authentication credentials in the
form of a username and password.

If a NULL password is specified the following protocol is used
for selecting the password to be used.

The TSEM_ES_PWD environment variable is checked.  If it is
defined the value of the environment variable is used as the
password.

If the environment variable is not set the following file is
checked for its existence, and if found, will use the contents of
that file as the password:

/opt/Quixote/etc/es.pwd

The initialization function also establishes a null callback
function in order to suppress output of any response from an ES
command.

The ->inject method takes a String object that contains a JSON
encoded string that specifies a document to be added to the index
that the object has been initialized to handle.