Implement tool for injecting documents into Elasticsearch.
The quixote-inject utility was built on top of the ESclient
object to provide a preliminary implementation of a tool for
injecting security event descriptions into an Elasticsearch
document.

The utility takes the following command-line arguments:

-h: Hostname of the Elasticsearch (ES) instance.

-i: The name of the index to add the documents to.

-u: The username to be used to authenticate to the ES instance.

The utility reads security event descriptions from standard
input, that are expected to be line feed terminated, and injects
the events into the specified index.