Use dedicated workload method for root export.
The TSEMworkload object was modified to remove the following
method:

->set_root_mode

And replace it with the following:

->run_root_export

The latter method now implements all of the functionality needed
to carry out the export of events from the root modeling
namespace.

The quixote-export utility was updated to drop the code that had
implemented the root export in favor of calling the new method.

The header documentation for the utility was updated and unused
and stale CPP defines were removed.