Implement boot identification for exported security descriptions.
The context key in the event{} description can be used to
differentite security event descriptions from different
implementations of the same security modeling namespace.  Input
from data analysis suggests that a data field is needed to
determine whether or not security event descriptions came from
different boots of an operating system.

This commit implements adding a 'boot identifier' to the export{}
field in order to allow this differentian.  The combination of
the boot identifier and the context number will now uniquely
establish the origin of the security event.

Linux, helpfully, supplies the necessary underlying technology
through the concept of a 'boot uuid'.  This value is uniquely
generated on each boot of the operating system and can be read
from the following pseudo-file:

/proc/sys/kernel/random/boot_id

The _read_boot_identifier() function was added as a helper
function to the main() function to encapsulate the reading of the
boot identifier.  The statically scoped Boot_ID String object is
used to hold that value for the duration of the invocation of the
utility.

The following function:

add_timestamp

Was renamed to the following:

update_event

Since there will now always be an update to the export field
generated.

The optional addition of a timestamp value was moved to the
following helper function function:

_add_timestamp

An export structure with the boot identifier and timestamp now
has the following form:

/proc/sys/kernel/random/boot_id
The context key in the event{} description can be used to
differentite security event descriptions from different
implementations of the same security modeling namespace.  Input
from data analysis suggests that a data field is needed to
determine whether or not security event descriptions came from
different boots of an operating system.

This commit implements adding a 'boot identifier' to the export{}
field in order to allow this differentian.  The combination of
the boot identifier and the context number will now uniquely
establish the origin of the security event.

Linux, helpfully, supplies the necessary underlying technology
through the concept of a 'boot uuid'.  This value is uniquely
generated on each boot of the operating system and can be read
from the following pseudo-file:

/proc/sys/kernel/random/boot_id

The _read_boot_identifier() function was added as a helper
function to the main() function to encapsulate the reading of the
boot identifier.  The statically scoped Boot_ID String object is
used to hold that value for the duration of the invocation of the
utility.

The following function:

add_timestamp

Was renamed to the following:

update_event

Since there will now always be an update to the export field
generated.

The optional addition of a timestamp value was moved to the
following helper function function:

_add_timestamp

"export": {"boot_id": "6911ca77-0ef5-4658-a0f8-af011fddaa50", "@timestamp": "2025-05-08T04:31:39.895252958", "type": "event"}