Add modeling capability to Elasticsearch injector.
This commit implements the ability to process security event
descriptions through a TSEM model before injecting the security
event description into a Quixote index.

The only security event descriptions that are injected are those
that have not yet occurred, or in the case of a preloaded model,
those events that are not consistent with the model.

The -M command-line switch was added to specify that a security
event description should be modeled before it is injected into an
index.

The -n command-line switch was added that disables the injection
of the security event description into a model.  The -v
command-line switch was added to print a security event
description that is injected.

In combination these two command-line switches allow a security
event trajectory to be screened and reduced to only the events
unique to a model.

The -m command-line option was added to allow a file containing a
security model definition to be loaded that the events are to be
processed against.