Add utility for generating a model definition.
This commit introduces the following utility:

generate-model

This utility reads a security trajectory log and generates a
model that describes the final security state of the system that
generated that trajectory.

A primary use of this tool is to generate a model definition from
a set of security descriptions that were generated by a security
modeling namespace running in export mode.