Handle event descriptions in compact form.
When index documents are read from Elasticsearch each document is
encoded as a member of an array.

The 'jq' utility can be used to extract the documents from the
array.  The utility has the ability to run in 'compact mode'
where each JSON encoded security event description is output with
no spaces that separate key/value pairs.

The JSON parsing in the TSEMparser object is a very simplistic
implementation that is designed to only parse security event
descriptions in the format that they are exported from the
kernel in, which have spaces between the key/value pairs.

This commit modifies the TSEMparser object to parse security
event descriptions that are in native kernel format as well as
compact format.