Prevent TSEM from generating segmentation faults.
A regression was noted where the 'generate-model' utility was
causing a segmentation fault when it processed a stream of
security event descriptions that generated all unique
coefficients.

The segmentation fault was being generated when the SecurityEvent
objects associated with the model were destroyed.

This was traced to the fact that the 'generate-model' utility was
calling the object destructor for the SecurityEvent object that
had been used to parse each incoming event.

Remove the destructor for that event and let all of the event
descriptions be cleaned up by the destruction of the model
object.