Implement cJSON build.
The TSEMparser object implements very limited JSON parsing that
is specific to the JSON that is generated by the Linux TSEM
kernel driver.  A more complete parser will be needed to parse
output from the Elasticsearch or Opensearch indexing engines.

The cJSON parser is an extremely minimal parser that consists of
only a single source and header file and will be considered for
use as a general purpose JSON parser for intereaction with
indexing engines.

Implement the download and build of the most recent version of
the cJSON parser.