Add output Buffer argument to ESclient->init method.
Implementing support for reading the documents out of an Elastic
or Open search index requires the ability to capture the output
of the Curl based request that is issued to interact with the
index instance.

This commit adds an additional argument in the form of a Buffer
object to the ESclient->init method.  The output from the Curl
command is accumulated in this object.

Passing a NULL pointer as the Buffer argument is used to indicate
that the command should run in 'quiet' mode with no output being
generated.  This is used, for example, with the quixote-inject
utility to suppress the output response from injecting a document
into the index.

This commit updates the quixote-index utility to use the new
signature for the ->init method.