Generalize index manipulation object.
Work to create a generic tool for managing document indexes
indicated the need to make the object more generic with respect
to how differing commands are implemented.

Three String objects were added to the object state to hold the
following data:

Endpoint for the command(s).

The name of the index.

Command construction object.

The ESclient->init object stores the portion of the URL that is
invariant across all commands, ie:

https://hostname:portnum

This string is used by subsequent commands to create the command
URL that is to be executed.

The syntax of the command that is to be executed is constructed
in each method that issues a command in the command construction
object.