Implement index list in ESclient.
This commit implements the ESclient->list method that queries a
search endpoint for all of the indexes that are supported by the
endpoint.

This implementation is the first capability in the ESclient
object that populates the output Buffer object.  A regression was
noted in that the Buffer object was not being populated with the
size of output that is being delivered to the output callback.
Only a single byte was being loaded into the buffer, this patch
updates the bufr->add method to use the size of the data that was
written to the callback.

This commit also fixes a regression in the ESclient->whack
method.  The WHACK macros was not being used that was causing a
segmentation fault since the ->list method was not initialized
the index object so it was a NULL pointer.  Replace the
explicitly coded destructors with the use of the WHACK macro.