Always output violating event descriptions.
When tracking a security event description stream the
quixote-inject utility should always output events that occur
that are in violation of a model being enforced.  This is so
that any adversarial events are always logged.

Gate the output of an event description either on a model update
or if the event was a violation of the security model.