Implement infrastructure for filesystem UUID substitution.
One of the issues associated with deterministic modeling is the
fact that various synthetic filesystems have randomly generated
filesystem UUID's.

Previously the only filesystem that was addressed was the TMPFS
filesystem but testing with runc based workloads indicated that
the following filesystem types can also have random UUID's:

sysfs
cgroup2

The static test for the TMPFS filesystem magic value was
subsituted with a call to the following helper function:

__substitute_uuid

This function loops over a set of filesystem magic numbers and
returns a boolean true value if the magic number presented to the
function matches on of the filesystem types.

If a boolean true value is returned by this helper function the
UUID is set to an all null byte value, as was done when the TMPFS
filesystem type was detected.