Implement JSON parser object.
Tooling which is being developed to allow security event
descriptions to be output from indexing engines requires a
general purpose JSON parser.  The TSEMparser object was designed
to be very specific to the JSON that is generated by the TSEM
kernel driver.

This commit implements the JSONparser object that is capable of
parsing general JSON syntax.  It is an object that supports a
wrapper around the cJSON library.

A unit test, the JSONparser_test binary, was added to test the
functioning of the parser.