Implement support for scrolling through an index.
By default the document index engines limit search output to
10,000 documents.  Since security trajectories can easily reach
into the millions of documents there is a need to support a
system that will allow more than 10,000 documents to be returned
in a single search request.

One of the prescribed ways to do that is through the 'scroll'
API.  A 'cookie' is returned in response to a request to initiate
a search that is used as a cursor to scroll through all of the
documents until the end of the index is reached.

This commit implements the following methods to support this
functionality:

ESclient->scroll
ESclient->next

The ->scroll method is responsible for initiating a search and
document dump.  This method takes a String object that defines
the search that is to be executed.  The Buffer object that has
been attached to the ESclient object is filled with the results
of the first search request, this is a JSON encoded document that
contains the scroll id, in addition to the first set of documents
that were requested.

The ->next method is called to request the next section of
documents.  It expects a String object containing the scroll idea
that will be used as the cursor for the next series of documents.

The Buffer object attached to the ESclient document is filled by
this method with a JSON encoded document containing the results
of that search and the next scroll identifier that is to be used.

This method implements only the mechanics of executing the
transaction with the indexing engine.  The parsing of the results
document and the extraction of the scroll identifier are the
responsibility of the caller.