Allow empty key values.
A regression was noted when processing a model based on Debian
root security events.

Some event descriptions for a socket_connect on an AF_UNIX socket
were returning with a 'NULL' value for the address key.  An
exampled JSON output is as follows:

{
  "export": {
    "boot_id": "dfedd6e8-c4ca-467a-8b49-89531aaa5f90",
    "@timestamp": "2025-06-30T18:41:54.625864339",
    "type": "event",
    "coefficient": "6c84fb0dca6836093460c3f055633fd6e428e2ef40f966649e884ff5df213ee5",
    "violation": "yes"
  },
  "event": {
    "pid": "1",
    "tnum": "1",
    "context": "0",
    "number": "5844720",
    "process": "systemd",
    "type": "socket_connect",
    "ttd": "79",
    "p_ttd": "0",
    "task_id": "f025bc32adc9ca4c01cabfe86bad82e8c465d98c2db24db5e18de3906e8a2953",
    "p_task_id": "0000000000000000000000000000000000000000000000000000000000000000",
    "ts": "4063355057482361"
  },
  "COE": {
    "uid": "0",
    "euid": "0",
    "suid": "0",
    "gid": "0",
    "egid": "0",
    "sgid": "0",
    "fsuid": "0",
    "fsgid": "0",
    "capeff": "0x3ffffffffff"
  },
  "socket_connect": {
    "sock": {
      "family": "1",
      "type": "1",
      "protocol": "0",
      "owner": "f025bc32adc9ca4c01cabfe86bad82e8c465d98c2db24db5e18de3906e8a2953"
    },
    "addr": {
      "af_unix": {
        "address": ""
      }
    }
  }
}

A null key value is valid JSON and the TSEMparser object should
handle this properly.

Check for this case and if it occurs reset the key value to the
null string.

This will cause the mesurement of the CELL value to not include
an address value in the input to the measurement value.