Implement dump mode for index utility.
This commit adds the ability to dump out the documents in an
index.  This function is selected by the -D command-line argument
to the qidx-util tool.

This update to the utility implements 'scroll' mode in order to
dump an index regardless of the number of documents in the index.
By default, searching for documents in an index is limited to
10,000 documents and it is not uncommon for Quixote event indexes
to have millions of security event descriptions in them.

The dumping of an index is encapsulated in the dump_index()
function.  This function uses the following helper functions to
implement the dumping:

_dump_array()
_scroll()

The dump_array() function creates a JSON document that specifies
the number of documents in a dump 'window' and the search
specification.  Since this is a 'dump' the search specification
is to match all of the documents in an index.

In response to processing of this document the indexing engine
returns a JSON document that defines the total number of
documents that match the search specification and the first
'window' of documents encoded as a JSON array of JSON objects.

The _dump_array() help function extracts the array of JSON
objects and prints the ASCII representation of each document.

The JSON response document contains a string value that is named
by the key name of '_scroll_id'.  This Base64 encoded string is a
cursor that is returned to the indexing engine to select the next
'window' of documents.

The total number of documents that match the search is decomposed
into the number of scroll cycles that are needed to dump all of
the documents that match the search specification.

The _scroll() helper function is called to encode the _scroll_id
in a JSON document that is submitted to the index engine to
select the next window of documents.  The _dump_array() function
is called by the _scroll() helper function to dump the array of
documents that is returned in response to the scroll request.